Move blog post inside blog
