+++ website/src/app/app-routing.module.ts
@@ -5,7 +5,7 @@ import { AboutComponent } from './ui/pages/about/about.component';
 import { BlogComponent } from './ui/pages/blog/blog.component';
 import { ContactComponent } from './ui/pages/contact/contact.component';
 import { HomeComponent } from './ui/pages/home/home.component';
-import { BlogPostComponent } from './ui/pages/blog-post/blog-post.component'
+import { BlogPostComponent } from './ui/pages/blog/blog-post/blog-post.component';
 
 const routes: Routes = [
     {path: '', component: HomeComponent},
