+++ website/src/app/app.module.ts
@@ -8,7 +8,7 @@ import { HomeComponent } from './ui/pages/home/home.component';
 import { AboutComponent } from './ui/pages/about/about.component';
 import { BlogComponent } from './ui/pages/blog/blog.component';
 import { ContactComponent } from './ui/pages/contact/contact.component';
-import { BlogPostComponent } from './ui/pages/blog-post/blog-post.component';
+import { BlogPostComponent } from './ui/pages/blog/blog-post/blog-post.component';
 
 @NgModule({
   declarations: [
