--- website/src/app/ui/pages/blog-post/blog-post.component.html
@@ -1,3 +1,0 @@
-<div>
-  <h1 class="title">{{post.title}}</h1>
-</div>
