+++ website/src/app/ui/pages/blog/blog-post/blog-post.component.html
@@ -1,0 +1,3 @@
+<div>
+  <!-- <h1 class="title">{{post.title}}</h1> -->
+</div>
